Mantis 994: o Fixed bug where the voice comm icon would not be displayed while using a mic
The new MemoryInputStream class either contains a bug or doesn't match what the vgui::BitmapTGA class expects. I've inserted the old code to get this working till Karl can take a look at it, as it's a bit painstaking to sort it out when I don't even know what method(s) it fails in.

I also fixed a bug in the constructor for the new MemoryInputStream where the length parameter would not be applied to the length member.

git-svn-id: https://unknownworlds.svn.cloudforge.com/ns1@232 67975925-1194-0748-b3d5-c16f83f1a3a1